фикс ошибок
