+++ app/src/main/java/com/example/tinkoffproject/di/NetworkModule.kt
@@ -1,6 +1,6 @@
 package com.example.tinkoffproject.di
 
-import com.example.tinkoffproject.BuildConfig
+import androidx.viewbinding.BuildConfig
 import com.example.tinkoffproject.data.FoodApi
 import com.example.tinkoffproject.data.interceptor.ApiKeyInterceptor
 import dagger.Module
@@ -14,7 +14,6 @@ import retrofit2.converter.gson.GsonConverterFactory
 import java.util.concurrent.TimeUnit
 
 private const val BASE_URL="https://api.spoonacular.com/"
-const val API_KEY="19dd87ea73ea4d618f7661859d85dda1"
 
 @Module
 @InstallIn(SingletonComponent::class)
