--- app/src/main/java/com/example/tinkoffproject/presentation/MainPagePackage/ProfileFragment.kt
@@ -1,4 +1,0 @@
-package com.example.tinkoffproject.presentation.MainPagePackage
-
-class ProfileFragment {
-}
