+++ app/src/main/java/com/example/tinkoffproject/presentation/mainPagePackage/FoodInfoFragment.kt
@@ -1,4 +1,4 @@
-package com.example.tinkoffproject.presentation.MainPagePackage
+package com.example.tinkoffproject.presentation.mainPagePackage
 
 import androidx.fragment.app.Fragment
