+++ app/src/main/java/com/example/tinkoffproject/presentation/mainPagePackage/ProfileFragment.kt
@@ -1,0 +1,4 @@
+package com.example.tinkoffproject.presentation.mainPagePackage
+
+class ProfileFragment {
+}
